project/requirements.txt
fastapi==0.65.2
uvicorn==0.14.0
moviepy==1.0.3
azure-cognitiveservices-speech==1.36.0
